perbaikilah ini :
Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
    at ShadowPortal (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/client/components/react-dev-overlay/ui/components/shadow-portal.tsx:5:32)
    at DevOverlay (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/client/components/react-dev-overlay/ui/dev-overlay.tsx:14:3)
    at ReactDevOverlay (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/server/dev/next-dev-server.ts:82:10)
    at div
    at Body (webpack://next/dist/src/server/render.tsx:1263:19)
Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
    at ShadowPortal (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/client/components/react-dev-overlay/ui/components/shadow-portal.tsx:5:32)
    at DevOverlay (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/client/components/react-dev-overlay/ui/dev-overlay.tsx:14:3)
    at ReactDevOverlay (/Users/wildanbelfiore/freelance_project/safarental-web/node_modules/next/src/server/dev/next-dev-server.ts:82:10)
    at div
    at Body (webpack://next/dist/src/server/render.tsx:1263:19)